Add the ability to filter tasks by authenticated user

+++ modules/tasks/TasksController.js
@@ -2,14 +2,17 @@ const Task = require('./Task')
 
 class TasksController {
   static async all (req, res) {
-    const tasks = await Task.query()
+    const tasks = await Task
+      .query()
+      .where('user_id', '=', req.user.sub)
     return res.send(tasks)
   }
 
   static async add (req, res) {
     const task = await Task.query().insert({
       'name': req.body['name'],
-      'category': req.body['category']
+      'category': req.body['category'],
+      'user_id': req.user.sub
     })
     return res.send(task)
   }
@@ -19,6 +22,7 @@ class TasksController {
       .query()
       .delete()
       .where('id', '=', req.body['id'])
+      .andWhere('user_id', '=', req.user.sub)
     return res.end()
   }
 }
